refactor: updte directories
